useComponent_simpleDir
