+++ src/App.js
@@ -1,20 +1,12 @@
 import logo from "./logo.svg";
 import "./App.css";
-import "./Board.css";
-import { useState } from "react";
+import "./page/Board.css";
+import BlackBoard from "./page/Blackboard";
 
 function App() {
-  const [NaeYong, CNaeYong] = useState("");
-
-  const PushText = (e) => {
-    CNaeYong(e.target.value);
-  };
-
   return (
     <div className="App">
-      <h1>디지털 칠판</h1>
-      <div className="MainText">{NaeYong}</div>
-      <input value={NaeYong} placeholder="입력해주세요" onChange={PushText} />
+      <BlackBoard />
     </div>
   );
 }
